[MNT] Isolate `torch` and `pyro` as soft dependencies (#2326)
Fixes #2325. Follows #2324 and #2266 - as @marrov noticed, `torch` is an
indirect dependency through `pyro-ppl`, hence in order to isolate
`torch`, we must also isolate `pyro-ppl` as a soft dependency.

This PR carries out isolation of `pyro-ppl` along the lines of #2324,
using the `_check_soft_dependencies` and `_safe_import` utilities from
`scikit-base`.

It also completes the isolation of `torch` as a soft dependency - due to
the indirect dependency on `pyro-ppl`, #2324 was incomplete.

Also fixes https://github.com/pgmpy/pgmpy/issues/2327 (this seems
necessary to avoid side effects on the global config, throughout the
tests)

---------

Co-authored-by: Ankur Ankan <ankurankan@gmail.com>